refactor: add table file to fix the lint error
